chore: remove constant files, declare inline
