--- packages/forge/src/lib/breadcrumbs/breadcrumbs-item/breadcrumbs-item-constants.ts
@@ -1,3 +1,0 @@
-import { COMPONENT_NAME_PREFIX } from '../../constants.js';
-
-export const BREADCRUMBS_ITEM_TAG_NAME: keyof HTMLElementTagNameMap = `${COMPONENT_NAME_PREFIX}breadcrumbs-item`;
