+++ packages/forge/src/lib/breadcrumbs/index.ts
@@ -4,9 +4,7 @@ import { BreadcrumbsComponent } from './breadcrumbs/breadcrumbs.js';
 import { BreadcrumbsItemComponent } from './breadcrumbs-item/breadcrumbs-item.js';
 
 export * from './breadcrumbs/breadcrumbs.js';
-export * from './breadcrumbs/breadcrumbs-constants.js';
 export * from './breadcrumbs-item/breadcrumbs-item.js';
-export * from './breadcrumbs-item/breadcrumbs-item-constants.js';
 
 /** @deprecated Definition functions are deprecated and replaced with side effect imports (`import '@tylertech/forge/breadcrumbs'`). */
 export function defineBreadcrumbsComponent(): void {
